Reimplement C++ Presolver using the PImpl pattern
This is a backwards compatibility break.

The immediate benefit is that we'll be much less likely to break
binary compatibility when we make implementation changes going
foward.
Another benefit is compilation time.
Finally, in the future, we can make the PresolverImpl class much
more testable because we can expose more methods without them
being user-facing.

Closes https://github.com/dwavesystems/dwave-preprocessing/issues/89